# Add try_unary, binary, try_binary kernels ~90% faster