… arrow/src/util/bit_iterator.rs